+++ kotlin-js-core/src/webMain/kotlin/js/core/JsPrimitives.kt
@@ -15,8 +15,8 @@ expect object JsPrimitives {
     inline fun JsByte.toByte(): Byte
     inline fun Byte.toJsByte(): JsByte
 
-    inline fun JsUByte.toUByte(): Short
-    inline fun Short.toJsUByte(): JsUByte
+    inline fun JsUByte.toUByte(): UByte
+    inline fun UByte.toJsUByte(): JsUByte
 
     inline fun JsShort.toShort(): Short
     inline fun Short.toJsShort(): JsShort
